Refactoring: extract class ListNode to leetCode/shared package
