+++ src/main/java/leetCode/mediumAlgorithms/listNode/AddTwoNumbers.java
@@ -1,6 +1,6 @@
-package leetCode.mediumAlgorithms.array;
+package leetCode.mediumAlgorithms.listNode;
 
-import java.util.Objects;
+import leetCode.shared.ListNode;
 
 /**
  * @author Anna S. Almielka
@@ -52,45 +52,6 @@ public class AddTwoNumbers {
             currentNode.next = new ListNode((dozen));
         }
         return result.next;
-    }
-
-}
-
-/**
- * Definition for singly-linked list.
- */
-class ListNode {
-    int val;
-    ListNode next;
-
-    ListNode() {
-    }
-
-    ListNode(int val) {
-        this.val = val;
-    }
-
-    ListNode(int val, ListNode next) {
-        this.val = val;
-        this.next = next;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-
-        ListNode listNode = (ListNode) o;
-
-        if (val != listNode.val) return false;
-        return Objects.equals(next, listNode.next);
-    }
-
-    @Override
-    public int hashCode() {
-        int result = val;
-        result = 31 * result + (next != null ? next.hashCode() : 0);
-        return result;
     }
 
 }
