+++ src/test/java/leetCodingChallenge/year2021/february/week1/Day03Test.java
@@ -1,5 +1,6 @@
 package leetCodingChallenge.year2021.february.week1;
 
+import leetCode.shared.ListNode;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
